[frontend] refactor restriction's Management component
