+++ opencti-platform/opencti-front/src/private/components/data/restriction/RestrictedDrafts.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const RestrictedDrafts = () => {
+  return <>hi</>;
+};
+
+export default RestrictedDrafts;
